+++ source/misc.h
@@ -9,7 +9,7 @@
 #define SOURCE_MISC_H_
 
 
-void Delay(uint16_t usec);
+void Delay(uint32_t usec);
 
 
 #endif /* SOURCE_MISC_H_ */
